+++ plugin-kit-react/src/utils/index.ts
@@ -4,6 +4,7 @@ export * from '@verbb/plugin-kit';
 // Export local utilities
 export * from './classes';
 export * from './config';
+export * from './documentScrollStability';
 export * from './craftHostBridge';
 export * from './handle';
 export * from './hostBridge';